Move to an Apple Wallet file.
